+++ packages/backup/mock_backupInterface_test.go
@@ -5,13 +5,10 @@ package backup
 import (
 	context "context"
 
+	v1 "github.com/cloudogu/k8s-backup-lib/api/v1"
 	mock "github.com/stretchr/testify/mock"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
 	types "k8s.io/apimachinery/pkg/types"
-
-	v1 "github.com/cloudogu/k8s-backup-lib/api/v1"
-
 	watch "k8s.io/apimachinery/pkg/watch"
 )